requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.3.9
cycler==0.10.0
Cython==0.29.10
decorator==4.4.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
imageio==2.5.0
imgaug==0.2.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
labelme==3.16.0
Markdown==3.1.1
matplotlib==3.1.0
mock==3.0.5
networkx==2.3
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.8.0
pydensecrf==1.0rc3
pygame==1.9.6
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1
QtPy==1.7.1
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Werkzeug==0.15.4
wincertstore==0.2
object-detection==0.1